Split follow sample into views and compose sub-packages
Moves the framework-specific screens into dedicated views/ and compose/
sub-packages so the two implementations are easy to find; shared files
(entry screen, synchronizer, view model, tabs configuration, data layer)
stay at the module root.
