+++ samples/follow/src/main/java/com/wscsports/blaze_sample_android/samples/follow/compose/FollowTabsComposeActivity.kt
@@ -1,4 +1,4 @@
-package com.wscsports.blaze_sample_android.samples.follow
+package com.wscsports.blaze_sample_android.samples.follow.compose
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
